Fix some shadow declarations and CodeQL warning (#156)
* Fix some shadow declarations and CodeQL warning

Possible fix for
https://github.com/netpanzer/netpanzer/security/code-scanning/30 related
to #104

+++ src/Lib/2D/Surface.hpp
@@ -58,9 +58,9 @@ class Surface : public NoCopy {
   void setOffsetCenter();
 
   // Accessor functions.
-  unsigned int getWidth() const { return twidth; }
-  unsigned int getHeight() const { return theight; }
-  unsigned int getPitch() const { return tpitch; }
+  unsigned short getWidth() const { return twidth; }
+  unsigned short getHeight() const { return theight; }
+  unsigned short int getPitch() const { return tpitch; }
 
   iXY getCenter() const { return iXY(getWidth() >> 1, getHeight() >> 1); }
   int getCenterX() const { return getWidth() >> 1; }
